Move custom rollup CSS bundler out to separate file.
Also update JSDocs to specify option parameters.
